web/container_data/requirements.txt
python-dotenv
flask
redis
pyjwt
requests
flask-cors